refactor: refactored the code
