+++ pkg/version/version.go
@@ -2,8 +2,6 @@
 
 package version
 
-// This file hosts a k8s suitable version object.
-
 import "k8s.io/apimachinery/pkg/version"
 
 // these are updated in build-time using LDFlags, note the Makefile
